requirements.txt
starlette==0.14.1
uvicorn
aiofiles
marshmallow
pillow
aiohttp

